+++ packages/skill-lesson/index.tsx
@@ -3,5 +3,4 @@ export * from './utils/build-stripe-checkout-path'
 export * from './trpc/routers/_skill-lesson-router'
 export * from './trpc/trpc-context'
 export * from './trpc/trpc.server'
-export * from './admin/coupon-generator'
 export * from './inngest/events'
