+++ packages/ui/index.tsx
@@ -1,1 +1,3 @@
 export * from './primitives'
+export * from './admin/coupon-data-table'
+export * from './admin/coupon-generator-form'
